Add project BallGame and finished rotation

+++ Bengine/IOManager.cpp
@@ -19,11 +19,11 @@ namespace Bengine
 		file.seekg(0, std::ios::end);
 
 		// Get the file size
-		int fileSize = file.tellg();
+		int fileSize = (int)file.tellg();
 		file.seekg(0, std::ios::beg);
 
 		// Reduce the file size by any heade3r bytes that might be present
-		fileSize -= file.tellg();
+		fileSize -= (int)file.tellg();
 
 		buffer.resize(fileSize);
 		file.read((char*)&(buffer[0]), fileSize);
